move converters into parsers namespace, drop unused functions
